Create sudokubot.py
Recode of an initial depth-first sudoku solver.  This version includes the depth-first solver as well as a solver based on a "fullest nine" logic, choosing the next cell based on which row, column, or cell of nine is the closest to being complete.  main() runs a speed comparison, depth-first x 10, fullest nines x 10, and then calculates average runtime for each approach.